requirements.txt
fastapi
uvicorn
fastembed
faiss-cpu
numpy
requests
beautifulsoup4
python-dotenv
groq
pydantic
gunicorn